Move move cmac parser to new parser class
