+++ wea/src/main/java/com/capstone/wea/controller/WEAController.java
@@ -4,21 +4,18 @@ import com.capstone.wea.model.cap.CAPMessageModel;
 import com.capstone.wea.model.cmac.*;
 import com.capstone.wea.model.sqlresult.*;
 import com.capstone.wea.model.sqlresult.mappers.*;
-import com.capstone.wea.processor.CAPProcessor;
+import com.capstone.wea.parser.XMLParser;
 import com.fasterxml.jackson.dataformat.xml.XmlMapper;
 import com.fasterxml.jackson.dataformat.xml.ser.ToXmlGenerator;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.server.ResponseStatusException;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
 import java.io.File;
 import java.net.URI;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
 
 @RequestMapping("/wea")
@@ -38,17 +35,7 @@ public class WEAController {
      */
     @GetMapping(value = "/getMessage", produces = "application/xml")
     public ResponseEntity<CMACMessageModel> getMessage() {
-        CMACMessageModel model = null;
-
-        try {
-            File message = new File("src/main/resources/sampleCmacMessage.xml");
-            XmlMapper mapper = new XmlMapper();
-            mapper.configure(ToXmlGenerator.Feature.WRITE_XML_DECLARATION, true);
-            model = mapper.readValue(message, CMACMessageModel.class);
-        } catch (Exception e) {
-            e.printStackTrace();
-            return ResponseEntity.internalServerError().build();
-        }
+        CMACMessageModel model = XMLParser.parseCMAC("src/main/resources/sampleCmacMessage.xml");
 
         return ResponseEntity.ok(model);
     }
@@ -174,7 +161,7 @@ public class WEAController {
      */
     @GetMapping(value = "/parseCapMessage", produces = "application/xml")
     public ResponseEntity<CAPMessageModel> parseCapMessage() {
-        CAPMessageModel result = CAPProcessor.parse("src/main/resources/sampleCapMessage.xml");
+        CAPMessageModel result = XMLParser.parseCAP("src/main/resources/sampleCapMessage.xml");
 
         //TODO: store converted message in database
         CMACMessageModel cmac = result.toCmac();
